% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliche_findr.R
\name{caliche_findr}
\alias{caliche_findr}
\title{A function to test if you've found a caliche layer}
\usage{
caliche_findr(state, elevation)
}
\arguments{
\item{state}{US state you are currently in}

\item{elevation}{Elevation at your location in feet}
}
\value{
Whether or not you have found a layer.
}
\description{
A function to test if you've found a caliche layer
}
