% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_directions.R
\name{helpful_directions}
\alias{helpful_directions}
\title{A function to help you get from here to there}
\usage{
helpful_directions(here, there)
}
\arguments{
\item{here}{Your starting location}

\item{there}{Your destination}
}
\value{
A snarky message ;)
}
\description{
A function to help you get from here to there
}
