% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_analysis.R
\name{pit_analysis}
\alias{pit_analysis}
\title{A function for viewing soil pit analysis}
\usage{
pit_analysis(layers, types, colors)
}
\arguments{
\item{layers}{The parameter for layer depths in cm}

\item{types}{soil texture as a string vector}

\item{colors}{Munsell colors as a string vector}
}
\value{

}
\description{
A function for viewing soil pit analysis
}
