% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grid7.R
\name{theme_grid7}
\alias{theme_grid7}
\title{Grid theme for ggplot2}
\usage{
theme_grid7()
}
\value{

}
\description{
A theme that highlights specific quadrants for your data, while helpfully deemphasizing your actual data.
}
